+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 - Implement `xsd:base64Binary` data type codec
 - Implement `xsd:hexBinary` data type codec
 - Implement `xsd:unsignedByte` data type codec
+- Implement `xsd:nonNegativeInteger` data type codec